Fixed StdTypes bug

+++ dev/Kernel/StdTypes.h
@@ -9,8 +9,8 @@ typedef unsigned char       UnsignedByte;
 typedef char                SignedByte;
 typedef unsigned short      UnsignedWord16;
 typedef short               SignedWord16;
-typedef unsigned long       UnsignedWord32;
-typedef long                SignedWord32;
+typedef unsigned int        UnsignedWord32;
+typedef int                 SignedWord32;
 typedef unsigned long long  UnsignedWord64;
 typedef long long           SignedWord64;
